+++ .eslintrc.js
@@ -1,7 +1,7 @@
 const { getESLintConfig } = require('@applint/spec');
 
 // https://www.npmjs.com/package/@iceworks/spec
-module.exports = getESLintConfig('react-ts', {
+module.exports = getESLintConfig('common-ts', {
   parserOptions: {
     project: [],
     createDefaultProgram: false,