+++ jest.config.js
@@ -11,12 +11,13 @@ module.exports = {
     '^.+\\.(j|t)sx?$': 'ts-jest',
   },
   roots: [
-    '<rootDir>/tests',
+    '<rootDir>',
   ],
   testPathIgnorePatterns: [
     '/node_modules/',
     '/lib/',
   ],
+  bail: 1,
   // copy from jest config
   testMatch: ['**/?(*.)+(spec|test).[jt]s?(x)'],
 };