+++ tests/fixture/keep_platform/kraken/input.js
@@ -14,4 +14,3 @@ if (isKraken) {
   console.log('others2');
 }
 
-