+++ tests/fixture/keep_platform/namespace/web/output.js
@@ -1,0 +1,11 @@
+var env = {
+    isWeb: true
+};
+
+if (env.isWeb) {
+    console.log('This is web');
+} else if (env.isWeex) {
+    console.log('This is weex');
+} else {
+    console.log('others');
+}