+++ tests/fixture/keep_platform/web/output.js
@@ -1,9 +1,8 @@
-var isWeex = false;
-var isWeb = true;
+var isWeb = true, isWeex = false;
 if (isWeb) {
-    console.log("This is web");
+    console.log('This is web');
 } else if (isWeex) {
-    console.log("This is weex");
+    console.log('This is weex');
 } else {
-    console.log("others");
+    console.log('others');
 }