+++ tests/fixture/remove_export_exprs/preserveConfigAndDefault/base/output.js
@@ -1,0 +1,5 @@
+export function getConfig() {
+  console.log(1);
+}
+
+export default class Home {}