+++ tests/fixture/remove_export_exprs/preserveData/remove-default-expr/input.js
@@ -1,0 +1,19 @@
+import fs from 'fs'
+import other from 'other'
+
+const [a, b, ...rest] = fs.promises
+const [foo, bar] = other
+
+export async function getData() {
+  a
+  b
+  rest
+  bar
+}
+
+export function getConfig() {
+}
+
+export default () => {
+  return <div />
+}