+++ tests/fixture/remove_export_exprs/preserveDefault/base/output.js
@@ -1,0 +1,7 @@
+import fs from 'fs';
+const [a] = fs.promises;
+export default class Home {
+    constructor(){
+        console.log(a);
+    }
+};