+++ tests/fixture/remove_export_exprs/preserveDefault/destructuring-assignment-array/output.js
@@ -1,0 +1,8 @@
+import fs from 'fs';
+import other from 'other';
+const [b] = fs.promises;
+const [foo] = other;
+export default function Home() {
+    console.log(b);
+    return __jsx("div", null);
+};