+++ tests/unminify/react/output.js
@@ -1,0 +1,13 @@
+function _classCallCheck(instance, Constructor) {
+    if (!(instance instanceof Constructor)) {
+        throw new TypeError("Cannot call a class as a function");
+    }
+}
+import React from "react";
+var Foo = function Foo() {
+    "use strict";
+    _classCallCheck(this, Foo);
+};
+export default function Home() {
+    return /*#__PURE__*/ React.createElement("div", null);
+};