+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## 0.2.0
+
+- chore: update swc to 0.95.0
+- feat: `options.keepPlatform` add import source @uni/env
+- fix: `options.keepPlatform`  analyse import namespace
+- feat: support remove exports
+
 ## 0.1.3
 
 - feat: remove multiple ends code